requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
databricks-sdk>=0.50
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
greenlet>=3.0.0
pyyaml>=6.0
alembic>=1.13.0

